chore(survey): move Area type to models
