+++ lib/surveys/view/select_area.dart
@@ -1,16 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:geolocator/geolocator.dart';
+import 'package:gibsonify_api/gibsonify_api.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:url_launcher/url_launcher.dart';
-
-class Area {
-  final LatLng center;
-  final int radius;
-  final double zoom;
-
-  Area(this.center, this.radius, this.zoom);
-}
 
 class SelectAreaScreen extends StatelessWidget {
   SelectAreaScreen({Key? key}) : super(key: key);
