+++ lib/surveys/view/view_area.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
-import 'package:gibsonify/surveys/surveys.dart';
+import 'package:gibsonify_api/gibsonify_api.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:url_launcher/url_launcher.dart';
