+++ packages/gibsonify_api/pubspec.yaml
@@ -14,6 +14,7 @@ dependencies:
   intl: ^0.17.0
   isar: ^3.0.5
   isar_flutter_libs: ^3.0.5
+  latlong2: ^0.8.1
 
 dev_dependencies:
   build_runner: ^2.3.2
